# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OtpRequestBody": {
        "properties": {
          "device_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "mobile": {
            "maxLength": 32,
            "minLength": 10,
            "title": "Mobile",
            "type": "string"
          }
        },
        "required": [
          "mobile"
        ],
        "title": "OtpRequestBody",
        "type": "object"
      },
      "OtpRequestResponse": {
        "properties": {
          "challenge_id": {
            "format": "uuid",
            "title": "Challenge Id",
            "type": "string"
          },
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          }
        },
        "required": [
          "challenge_id",
          "expires_in_seconds"
        ],
        "title": "OtpRequestResponse",
        "type": "object"
      },
      "OtpVerifyBody": {
        "properties": {
          "challenge_id": {
            "format": "uuid",
            "title": "Challenge Id",
            "type": "string"
          },
          "code": {
            "pattern": "^\\d{6}$",
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "code"
        ],
        "title": "OtpVerifyBody",
        "type": "object"
      },
      "OtpVerifyResponse": {
        "properties": {
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token"
          },
          "attempts_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempts Remaining"
          },
          "expires_in_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In Seconds"
          },
          "identity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identity Id"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          },
          "state": {
            "enum": [
              "verified",
              "invalid",
              "expired",
              "consumed",
              "locked",
              "not_found"
            ],
            "title": "State",
            "type": "string"
          },
          "token_type": {
            "anyOf": [
              {
                "const": "bearer",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Type"
          }
        },
        "required": [
          "state"
        ],
        "title": "OtpVerifyResponse",
        "type": "object"
      },
...
      "RefreshBody": {
        "properties": {
          "refresh_token": {
            "maxLength": 512,
            "minLength": 32,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshBody",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in_seconds"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Pet Platform Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/otp/request": {
      "post": {
        "operationId": "request_otp_api_v1_auth_otp_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Otp",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "operationId": "verify_otp_api_v1_auth_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpVerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Otp",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/session/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_session_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/session/refresh": {
      "post": {
        "operationId": "refresh_session_api_v1_auth_session_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Session",
        "tags": [
          "authentication"
        ]
      }
    },
...
  }
}